Added export/import test

+++ deployment/initialize_keycloak.sh
@@ -51,6 +51,19 @@ timeout 300 bash -c "while ! ssh ubuntu@\"$target_server_url\" \"cd $location &&
                        sleep 5;
                      done"
 
+ssh ubuntu@"$target_server_url" "sudo docker exec $container_name /opt/keycloak/bin/kcadm.sh "
+
+echo "Testing if the number of current users matches the number of exported users"
+current_users=$(ssh ubuntu@"$target_server_url" "docker exec $container_name /opt/keycloak/bin/kcadm.sh get users -r datalandsecurity --server http://localhost:8080/keycloak --realm master --user $KEYCLOAK_ADMIN --password $KEYCLOAK_ADMIN_PASSWORD | grep -c '\"username\" :'")
+all_users=$(ssh ubuntu@"$target_server_url" "docker exec $container_name ls /keycloak_users/datalandsecurity-users-*.json | wc -l")
+test_users=$(ssh ubuntu@"$target_server_url" "docker exec $container_name grep -E -l \"test_user.*@dataland.com\" /keycloak_users/datalandsecurity-users-*.json | wc -l")
+expected_users=$((all_users-test_users))
+if [[ ! $expected_users -eq $current_users ]]; then
+  echo "Found $current_users but $expected_users were expected."
+  exit 1
+fi
+echo "Imported users match the exported users."
+
 echo "Shutting down all running containers."
 ssh ubuntu@"$target_server_url" 'sudo docker kill $(sudo docker ps -q); sudo docker system prune --force; sudo docker info'
